requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
cloudpickle==2.2.1
cmake==3.25.0
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
decorator==4.4.2
filelock==3.9.0
fonttools==4.42.1
fsspec==2023.4.0
gym==0.26.2
gymnasium==1.1.1
gym-notices==0.0.8
gym-super-mario-bros==7.4.0
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
Jinja2==3.1.2
kiwisolver==1.4.5
lit==15.0.7
lz4==4.3.2
MarkupSafe==2.1.2
matplotlib==3.7.2
moviepy==1.0.3
mpmath==1.2.1
nes-py==8.2.1
networkx==3.0
numpy==1.23.5
opencv-python==4.8.0.76
packaging==23.1
Pillow==9.3.0
proglog==0.1.10
pyglet==1.5.21
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.1
six==1.16.0
sympy==1.11.1
tensordict==0.2.1
# torch==2.1.1+cu121
# torchaudio==2.1.1+cu121
torchrl==0.2.1
# torchvision==0.16.1+cu121
tqdm==4.66.1
typing_extensions==4.4.0
urllib3==1.26.13
